Ported from 5.1 to 5.3

+++ Source/Game1812/Actors/PiecesSpawner.cpp
@@ -17,7 +17,7 @@ void APiecesSpawner::BeginPlay()
 {
 	Super::BeginPlay();
 
-	const int SideLength = FMath::CeilToInt(FMath::Sqrt(Amount));
+	const int SideLength = FMath::CeilToInt(FMath::Sqrt((float)Amount));
 
 	for (int x = 0, i = 0; x < SideLength; x++) 
 	{
